requirements/base.txt
django-environ==0.11.2
kombu==5.3.4
fsmpy==2.1.0
peewee==3.17.0
sentry-sdk==2.3.1
